requirements.txt
osmnx
opencv-python
mercantile
wget
requests
tqdm
Pillow
open_clip_torch
gpustat